ui/saved_objects components 👉 New Platform (#43416)
* refactor: 💡 remove chrome usage in ui/saved_objects

* refactor: 💡 move <SavedObjectFinder> to kibana-react NP plugin

* test: 💍 move <SavedObjectFinder> tests to New Platform

* refactor: 💡 move <SavedObjectSaveModal> to kibana-react plugin

* test: 💍 fix TypeScript errors in a test

* feat: 🎸 port translations to NP plugin

* test: 💍 update Jest snapshot
