+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -506,24 +506,24 @@
     "common.ui.savedObjects.confirmModal.overwriteTitle": "{name} を上書きしますか？",
     "common.ui.savedObjects.confirmModal.saveDuplicateButtonLabel": "{name} を保存",
     "common.ui.savedObjects.confirmModal.saveDuplicateConfirmationMessage": "「{title}」というタイトルの {name} が既に存在します。保存を続けますか？",
-    "common.ui.savedObjects.finder.filterButtonLabel": "タイプ",
-    "common.ui.savedObjects.finder.searchPlaceholder": "検索…",
-    "common.ui.savedObjects.finder.sortAsc": "昇順",
-    "common.ui.savedObjects.finder.sortAuto": "ベストマッチ",
-    "common.ui.savedObjects.finder.sortButtonLabel": "並べ替え",
-    "common.ui.savedObjects.finder.sortDesc": "降順",
+    "kibana-react.savedObjects.finder.filterButtonLabel": "タイプ",
+    "kibana-react.savedObjects.finder.searchPlaceholder": "検索…",
+    "kibana-react.savedObjects.finder.sortAsc": "昇順",
+    "kibana-react.savedObjects.finder.sortAuto": "ベストマッチ",
+    "kibana-react.savedObjects.finder.sortButtonLabel": "並べ替え",
+    "kibana-react.savedObjects.finder.sortDesc": "降順",
     "common.ui.savedObjects.howToSaveAsNewDescription": "Kibana の以前のバージョンでは、{savedObjectName} の名前を変更すると新しい名前でコピーが作成されました。今後この操作を行うには、「新規 {savedObjectName} として保存」を使用します。",
     "common.ui.savedObjects.overwriteRejectedDescription": "上書き確認が拒否されました",
     "common.ui.savedObjects.saveAsNewLabel": "新規 {savedObjectName} として保存",
     "common.ui.savedObjects.saveDuplicateRejectedDescription": "重複ファイルの保存確認が拒否されました",
-    "common.ui.savedObjects.saveModal.cancelButtonLabel": "キャンセル",
-    "common.ui.savedObjects.saveModal.confirmSaveButtonLabel": "保存の確認",
-    "common.ui.savedObjects.saveModal.duplicateTitleDescription": "重複タイトルで {objectType} を保存するには {confirmSaveLabel} をクリックしてください。",
-    "common.ui.savedObjects.saveModal.duplicateTitleDescription.confirmSaveText": "保存の確認",
-    "common.ui.savedObjects.saveModal.duplicateTitleLabel": "「{title}」というタイトルの {objectType} が既に存在します。",
-    "common.ui.savedObjects.saveModal.saveAsNewLabel": "新規 {objectType} として保存",
-    "common.ui.savedObjects.saveModal.saveTitle": "{objectType} を保存",
-    "common.ui.savedObjects.saveModal.titleLabel": "タイトル",
+    "kibana-react.savedObjects.saveModal.cancelButtonLabel": "キャンセル",
+    "kibana-react.savedObjects.saveModal.confirmSaveButtonLabel": "保存の確認",
+    "kibana-react.savedObjects.saveModal.duplicateTitleDescription": "重複タイトルで {objectType} を保存するには {confirmSaveLabel} をクリックしてください。",
+    "kibana-react.savedObjects.saveModal.duplicateTitleDescription.confirmSaveText": "保存の確認",
+    "kibana-react.savedObjects.saveModal.duplicateTitleLabel": "「{title}」というタイトルの {objectType} が既に存在します。",
+    "kibana-react.savedObjects.saveModal.saveAsNewLabel": "新規 {objectType} として保存",
+    "kibana-react.savedObjects.saveModal.saveTitle": "{objectType} を保存",
+    "kibana-react.savedObjects.saveModal.titleLabel": "タイトル",
     "common.ui.scriptingLanguages.errorFetchingToastDescription": "Elasticsearch から利用可能なスクリプト言語の取得中にエラーが発生しました",
     "common.ui.share.contextMenu.embedCodeLabel": "埋め込みコード",
     "common.ui.share.contextMenu.embedCodePanelTitle": "埋め込みコード",
