+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -506,24 +506,24 @@
     "common.ui.savedObjects.confirmModal.overwriteTitle": "覆盖“{name}”？",
     "common.ui.savedObjects.confirmModal.saveDuplicateButtonLabel": "保存“{name}”",
     "common.ui.savedObjects.confirmModal.saveDuplicateConfirmationMessage": "具有标题 “{title}” 的 “{name}” 已存在。是否确定要保存？",
-    "common.ui.savedObjects.finder.filterButtonLabel": "类型",
-    "common.ui.savedObjects.finder.searchPlaceholder": "搜索……",
-    "common.ui.savedObjects.finder.sortAsc": "升序",
-    "common.ui.savedObjects.finder.sortAuto": "最佳匹配",
-    "common.ui.savedObjects.finder.sortButtonLabel": "排序",
-    "common.ui.savedObjects.finder.sortDesc": "降序",
+    "kibana-react.savedObjects.finder.filterButtonLabel": "类型",
+    "kibana-react.savedObjects.finder.searchPlaceholder": "搜索……",
+    "kibana-react.savedObjects.finder.sortAsc": "升序",
+    "kibana-react.savedObjects.finder.sortAuto": "最佳匹配",
+    "kibana-react.savedObjects.finder.sortButtonLabel": "排序",
+    "kibana-react.savedObjects.finder.sortDesc": "降序",
     "common.ui.savedObjects.howToSaveAsNewDescription": "在 Kibana 的以前版本中，更改 {savedObjectName} 的名称将创建具有新名称的副本。使用“另存为新的 {savedObjectName}” 复选框可立即达到此目的。",
     "common.ui.savedObjects.overwriteRejectedDescription": "已拒绝覆盖确认",
     "common.ui.savedObjects.saveAsNewLabel": "另存为新的 {savedObjectName}",
     "common.ui.savedObjects.saveDuplicateRejectedDescription": "已拒绝使用重复标题保存确认",
-    "common.ui.savedObjects.saveModal.cancelButtonLabel": "取消",
-    "common.ui.savedObjects.saveModal.confirmSaveButtonLabel": "确认保存",
-    "common.ui.savedObjects.saveModal.duplicateTitleDescription": "单击 “{confirmSaveLabel}” 以保存标题重复的{objectType}",
-    "common.ui.savedObjects.saveModal.duplicateTitleDescription.confirmSaveText": "确认保存",
-    "common.ui.savedObjects.saveModal.duplicateTitleLabel": "具有标题 “{title}” 的 {objectType} 已存在。",
-    "common.ui.savedObjects.saveModal.saveAsNewLabel": "另存为新的{objectType}",
-    "common.ui.savedObjects.saveModal.saveTitle": "保存{objectType}",
-    "common.ui.savedObjects.saveModal.titleLabel": "标题",
+    "kibana-react.savedObjects.saveModal.cancelButtonLabel": "取消",
+    "kibana-react.savedObjects.saveModal.confirmSaveButtonLabel": "确认保存",
+    "kibana-react.savedObjects.saveModal.duplicateTitleDescription": "单击 “{confirmSaveLabel}” 以保存标题重复的{objectType}",
+    "kibana-react.savedObjects.saveModal.duplicateTitleDescription.confirmSaveText": "确认保存",
+    "kibana-react.savedObjects.saveModal.duplicateTitleLabel": "具有标题 “{title}” 的 {objectType} 已存在。",
+    "kibana-react.savedObjects.saveModal.saveAsNewLabel": "另存为新的{objectType}",
+    "kibana-react.savedObjects.saveModal.saveTitle": "保存{objectType}",
+    "kibana-react.savedObjects.saveModal.titleLabel": "标题",
     "common.ui.scriptingLanguages.errorFetchingToastDescription": "从 Elasticsearch 获取可用的脚本语言时出错",
     "common.ui.share.contextMenu.embedCodeLabel": "嵌入代码",
     "common.ui.share.contextMenu.embedCodePanelTitle": "嵌入代码",
